Confirmando a exclusão e exibindo um Toast

+++ bluetasks-frontend/src/components/TaskListTable.js
@@ -1,5 +1,7 @@
 import React, { Component } from "react";
 import TaskService from "../api/TaskService";
+import { toast, ToastContainer } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 class TaskListTable extends Component {
   constructor(props) {
@@ -21,16 +23,25 @@ class TaskListTable extends Component {
   }
 
   onDeleteHandler(id) {
-    TaskService.delete(id);
-    this.listTasks();
+    if (window.confirm(`Deseja mesmo excluir a Tarefa ${id}?`)) {
+      TaskService.delete(id);
+      this.listTasks();
+      toast.success(`Tarefa ${id} excluída!`, {
+        position: toast.POSITION.BOTTOM_LEFT,
+      });
+    }
   }
 
   render() {
     return (
-      <table className="table table-striped">
-        <TableHeader />
-        <TableBody tasks={this.state.tasks} onDelete={this.onDeleteHandler} />
-      </table>
+      <>
+        <table className="table table-striped">
+          <TableHeader />
+          <TableBody tasks={this.state.tasks} onDelete={this.onDeleteHandler} />
+        </table>
+
+        <ToastContainer autoClose={2000} />
+      </>
     );
   }
 }
